Renames Period to TrainingPlan and optimizes imports
