+++ src/main/kotlin/it/unibo/lss/fcla/athleticpreparation/domain/model/TrainingPlan.kt
@@ -3,5 +3,5 @@ package it.unibo.lss.fcla.athleticpreparation.domain.model
 /**
  * @author Nicola Lasagni on 22/02/2021.
  */
-class Period {
+class TrainingPlan {
 }
